Add initial networking setup
